# frontend/openapi/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v2/actorTypes:
    post:
      operationId: ActorTypeController_create
      summary: Créer un nouveau type d’acteur
      description: "

        **Ce endpoint permet de créer un type d’acteur complète**


        Vous devez fournir les informations suivantes :

        - **code**: Le code du type d’acteur

        - **label**: Le libellé du type d’acteur

        - **description**: La description du type d’acteur

        \    "
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActorTypeDto"
      responses:
        "201":
          description: type d’acteur créé avec succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorTypeDto"
      tags: &a14
        - actorTypes
    get:
      operationId: ActorTypeController_findAll
      summary: Récupérer tous les types d’acteurs
      description: Ce endpoint permet de récupérer la liste de tous les types
        d’acteurs disponibles.
      parameters:
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
      responses:
        "200":
          description: Liste les types d’acteurs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActorTypeDto"
      tags: *a14
  /api/v2/actorTypes/perms-matrix:
    get:
      operationId: ActorTypeController_getMatrix
      summary: Récupérer la matrice des permissions
      description: Cet endpoint permet de récupérer la matrice des permissions pour
        les types d’acteurs.
      parameters: []
      responses:
        "200":
          description: Liste les types d’acteurs par id et de leurs permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppPermsDto"
      tags: *a14
    patch:
      operationId: ActorTypeController_updateMatrix
      summary: Mettre à jour la matrice des permissions
      description: Cet endpoint permet de mettre à jour la matrice des permissions
        pour les types d’acteurs.
      parameters: []
      requestBody:
        required: true
        description: Liste des permissions à mettre à jour pour les types d’acteurs
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AppPermsDto"
      responses:
        "200":
          description: Met à jour la matrice des permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppPermsDto"
      tags: *a14
  /api/v2/actorTypes/{id}:
    get:
      operationId: ActorTypeController_findOne
      summary: Récupérer un type d’acteur spécifique par ID
      description: Cet endpoint permet de récupérer les détails complets d'un type
        d’acteur en fonction de son identifiant unique.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Récupère un type d’acteur spécifique par ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorTypeDto"
      tags: *a14
    patch:
      operationId: ActorTypeController_update
      summary: Mettre à jour un type d’acteur
      description: "

        Ce endpoint permet de mettre à jour un type d’acteur existante

        Vous devez fournir l'identifiant du type d’acteur dans l'URL et les
        nouvelles données dans le corps de la requête

        Les données de mise à jour doivent correspondre aux champs

        \    "
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchActorTypeDto"
      responses:
        "200":
          description: Type d’acteur mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchActorTypeDto"
      tags: *a14
    delete:
      operationId: ActorTypeController_delete
      summary: Supprimer un type d’acteur
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Type d’acteur supprimé avec succès
      tags: *a14
  /api/v2/applications/{applicationId}/actors:
    post:
      operationId: ApplicationActorsController_create
      summary: Créer un nouvel acteur
      description: "

        Ce endpoint permet de créer un acteur complet.


        Informations requises :\ 

        - **email** : Email de l'acteur

        - **firstname** : Prénom de l'acteur

        - **lastname** : Nom de l'acteur

        - **actorTypeId** : ID du type d'acteur lié à l'acteur\ 

        - **organizationId** : ID de l'organisation liée à l'acteur

        - **applicationId** : ID de l'application liée à l'acteur

        \    "
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActorDto"
      responses:
        "201":
          description: Acteur créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorDto"
      tags: &a15
        - Actors
    get:
      operationId: ApplicationActorsController_findAll
      summary: Récupérer tous les acteurs
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
      responses:
        "200":
          description: Liste des acteurs trouvés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActorDto"
      tags: *a15
  /api/v2/applications/{applicationId}/actors/{id}:
    get:
      operationId: ApplicationActorsController_findOne
      summary: Récupérer un acteur par ID
      parameters:
        - name: id
          required: true
          in: path
          description: ID de l'acteur
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      responses:
        "200":
          description: Acteur trouvé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorDto"
      tags: *a15
    patch:
      operationId: ApplicationActorsController_updated
      summary: Mettre à jour un acteur
      parameters:
        - name: id
          required: true
          in: path
          description: ID de l'acteur
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActorDto"
      responses:
        "200":
          description: Acteur mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorDto"
      tags: *a15
    delete:
      operationId: ApplicationActorsController_delete
      summary: Supprimer un acteur
      parameters:
        - name: id
          required: true
          in: path
          description: ID de l'acteur
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      responses:
        "204":
          description: Acteur supprimé avec succès
      tags: *a15
...
info:
  title: API Référentiel Applications
  description: API pour la gestion des applications
  version: "2.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/auth
          tokenUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          refreshUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          scopes:
            openid: OpenID scope
            profile: Profile scope
      description: OAuth2 authentication using OIDC
    api_key:
      type: apiKey
      in: header
      name: x-refapp-token
      description: Token authentication
  schemas:
...
    CreateActorTypeDto:
      type: object
      properties:
        code:
          type: string
          example: MOA
          description: Code du type d'acteur
        label:
          type: string
          example: Maîtrise d’Ouvrage
          description: Libellé du type d'acteur
        description:
          type: string
          example: La MOA définit les besoins du projet, supervise sa mise en œuvre et
            veille à ce que les objectifs métiers soient atteints.
          description: description du type d'acteur
      required:
        - code
        - label
    ActorTypeDto:
      type: object
      properties:
        code:
          type: string
          example: MOA
          description: Code du type d'acteur
        label:
          type: string
          example: Maîtrise d’Ouvrage
          description: Libellé du type d'acteur
        description:
          type: string
          example: La MOA définit les besoins du projet, supervise sa mise en œuvre et
            veille à ce que les objectifs métiers soient atteints.
          description: description du type d'acteur
        id:
          type: string
          example: 12345678-1234-1234-1234-123456789012
          description: Identifiant unique du type d'acteur
      required:
        - code
        - label
        - id
    AppPermsDto:
      type: object
      properties:
        actorTypeId:
          type: string
        readBase:
          type: boolean
        writeBase:
          type: boolean
        writePriorityRestart:
          type: boolean
        readActors:
          type: boolean
        writeActors:
          type: boolean
        readCompliances:
          type: boolean
        writeCompliances:
          type: boolean
        readHostings:
          type: boolean
        writeHostings:
          type: boolean
        readMetadata:
          type: boolean
        readRelations:
          type: boolean
        writeRelations:
          type: boolean
        readLinks:
          type: boolean
        writeLinks:
          type: boolean
        readReports:
          type: boolean
        postReports:
          type: boolean
        manageReports:
          type: boolean
      required:
        - actorTypeId
        - readBase
        - writeBase
        - writePriorityRestart
        - readActors
        - writeActors
        - readCompliances
        - writeCompliances
        - readHostings
        - writeHostings
        - readMetadata
        - readRelations
        - writeRelations
        - readLinks
        - writeLinks
        - readReports
        - postReports
        - manageReports
    PaginatedActorTypeDto:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/ActorTypeDto"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    PatchActorTypeDto:
      type: object
      properties:
        code:
          type: string
          example: MOA
          description: Code du type d'acteur
        label:
          type: string
          example: Maîtrise d’Ouvrage
          description: Libellé du type d'acteur
        description:
          type: string
          example: La MOA définit les besoins du projet, supervise sa mise en œuvre et
            veille à ce que les objectifs métiers soient atteints.
          description: description du type d'acteur
    CreateActorDto:
      type: object
      properties:
        email:
          type: string
          example: example@example.com
          description: Email de l'acteur (Optionel)
        firstname:
          type: string
          example: Jean
          description: Prénom de l’acteur
        lastname:
          type: string
          example: Dupont
          description: Nom de l’acteur
        actorTypeId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID du type d'acteur lié
        organizationId:
          type: string
          nullable: true
          example: 9965dc19-1c5a-472d-83e3-8bf65093c86c
          description: ID de l'organization lié (Optionel)
        applicationId:
          type: string
          nullable: true
          example: 035869dc-47a6-4cee-828f-f6a28d050b35
          description: ID de l'application lié (Optionel)
      required:
        - actorTypeId
    ActorDto:
      type: object
      properties:
        email:
          type: string
          example: example@example.com
          description: Email de l'acteur (Optionel)
        firstname:
          type: string
          example: Jean
          description: Prénom de l’acteur
        lastname:
          type: string
          example: Dupont
          description: Nom de l’acteur
        actorTypeId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID du type d'acteur lié
        organizationId:
          type: string
          nullable: true
          example: 9965dc19-1c5a-472d-83e3-8bf65093c86c
          description: ID de l'organization lié (Optionel)
        applicationId:
          type: string
          nullable: true
          example: 035869dc-47a6-4cee-828f-f6a28d050b35
          description: ID de l'application lié (Optionel)
        id:
          type: string
          example: 12345678-1234-1234-1234-123456789012
          description: Identifiant unique de l'acteur
      required:
        - actorTypeId
        - id
    PaginatedActorDto:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/ActorDto"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    UpdateActorDto:
      type: object
      properties:
        email:
          type: string
          example: example@example.com
          description: Email de l'acteur (Optionel)
        firstname:
          type: string
          example: Jean
          description: Prénom de l’acteur
        lastname:
          type: string
          example: Dupont
          description: Nom de l’acteur
        actorTypeId:
          type: string
          example: 5708d232-8338-4abf-8f38-8370acc89497
          description: ID du type d'acteur lié
        organizationId:
          type: string
          nullable: true
          example: 9965dc19-1c5a-472d-83e3-8bf65093c86c
          description: ID de l'organization lié (Optionel)
        applicationId:
          type: string
          nullable: true
          example: 035869dc-47a6-4cee-828f-f6a28d050b35
          description: ID de l'application lié (Optionel)
...
security:
  - oauth2: []
  - api_key: []